# specs/001-slice-1-core/contracts/control-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Understudy Control API",
    "version": "0.1.0",
    "description": "Control surface for a running understudy mock. Every capability of the tool is reachable here; the CLI is a client of this API and adds no logic of its own.\n",
    "x-served-source": "This document is a CHECKED-IN artefact, not generated from the implementation. The running control plane serves it byte-identical; a test asserts that (a drift check). If a served description were generated from the code it could never disagree with the implementation, which is the opposite of a contract.\n",
    "x-prefix": "Every path below is relative to the configured reserved prefix (default /__understudy). The prefix is composed as prefix + path and carries no trailing slash.\n"
  },
  "paths": {
...
    "/requests": {
      "get": {
        "operationId": "listRequests",
        "summary": "The request log",
        "parameters": [
          {
            "name": "method",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "live",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching requests, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "requests",
                    "total"
                  ],
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "requests": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RequestLogEntry"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MalformedRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "MalformedRequest": {
        "description": "The request body or parameters could not be read",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ControlError"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The control operation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ControlError"
            }
          }
        }
      }
    },
    "schemas": {
      "ControlError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A stable short code, e.g. malformed_request or unknown_control_operation."
          },
          "message": {
            "type": "string",
            "description": "Human-readable, and names the offending field or path."
          },
          "field": {
            "type": "string",
            "description": "The offending input, when the error is attributable to one."
          }
        }
      },
...
      "RequestLogEntry": {
        "type": "object",
        "required": [
          "at",
          "method",
          "path",
          "status",
          "live"
        ],
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "live": {
            "type": "boolean"
          },
          "durationMs": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}